requirements.txt
kivy
kivymd
pillow == 8.4.0
packaging
plyer
